fix some bugs

+++ db/migrate/20220516110947_create_connections.rb
@@ -2,7 +2,7 @@ class CreateConnections < ActiveRecord::Migration[6.1]
   def change
     create_table :connections do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer :remote_id
+      t.bigint :remote_id
       t.boolean :arhived, default: false
 
       t.timestamps null: false
